+++ src/App.tsx
@@ -4,12 +4,16 @@ import React, { useEffect, useState } from 'react';
 import { UserWarning } from './UserWarning';
 import { USER_ID, getTodos } from './api/todos';
 import { Todo } from './types/Todo';
+import { FilterStatus } from './types/FilterStatus';
+import { TodoFilter } from './components/TodoFilter';
+import { TodoList } from './components/TodoList';
+import { ErrorNotification } from './components/ErrorNotification';
 
 export const App: React.FC = () => {
   const [todos, setTodos] = useState<Todo[]>([]);
   const [isLoading, setIsLoading] = useState(false);
   const [error, setError] = useState('');
-  const [filterStatus, setFilterStatus] = useState('all');
+  const [filterStatus, setFilterStatus] = useState(FilterStatus.All);
 
   useEffect(() => {
     const loadTodos = async () => {
@@ -48,22 +52,6 @@ export const App: React.FC = () => {
     setError('');
   };
 
-  const filteredTodos = todos.filter(todo => {
-    if (filterStatus === 'all') {
-      return true;
-    }
-
-    if (filterStatus === 'active') {
-      return !todo.completed;
-    }
-
-    if (filterStatus === 'completed') {
-      return todo.completed;
-    }
-
-    return true;
-  });
-
   const activeTodosCount = todos.filter(todo => !todo.completed).length;
   const hasCompletedTodos = todos.some(todo => todo.completed);
   const areAllCompleted =
@@ -97,51 +85,11 @@ export const App: React.FC = () => {
           </form>
         </header>
 
-        <section className="todoapp__main" data-cy="TodoList">
-          {isLoading && !todos.length ? (
-            <div className="modal overlay is-active">
-              <div className="modal-background has-background-white-ter" />
-              <div className="loader" />
-            </div>
-          ) : (
-            filteredTodos.map(todo => (
-              <div
-                key={todo.id}
-                data-cy="Todo"
-                className={`todo ${todo.completed ? 'completed' : ''}`}
-              >
-                <label className="todo__status-label">
-                  <input
-                    data-cy="TodoStatus"
-                    type="checkbox"
-                    className="todo__status"
-                    checked={todo.completed}
-                    readOnly
-                  />
-                </label>
-
-                <span data-cy="TodoTitle" className="todo__title">
-                  {todo.title}
-                </span>
-
-                {/* Remove button appears only on hover */}
-                <button
-                  type="button"
-                  className="todo__remove"
-                  data-cy="TodoDelete"
-                >
-                  ×
-                </button>
-
-                {/* overlay will cover the todo while it is being deleted or updated */}
-                <div data-cy="TodoLoader" className="modal overlay">
-                  <div className="modal-background has-background-white-ter" />
-                  <div className="loader" />
-                </div>
-              </div>
-            ))
-          )}
-        </section>
+        <TodoList
+          todos={todos}
+          filterStatus={filterStatus}
+          isLoading={isLoading}
+        />
 
         {/* Hide the footer if there are no todos */}
         {todos.length > 0 && (
@@ -151,43 +99,10 @@ export const App: React.FC = () => {
             </span>
 
             {/* Active link should have the 'selected' class */}
-            <nav className="filter" data-cy="Filter">
-              <a
-                href="#/"
-                className={`filter__link ${filterStatus === 'all' ? 'selected' : ''}`}
-                data-cy="FilterLinkAll"
-                onClick={e => {
-                  e.preventDefault();
-                  setFilterStatus('all');
-                }}
-              >
-                All
-              </a>
-
-              <a
-                href="#/active"
-                className={`filter__link ${filterStatus === 'active' ? 'selected' : ''}`}
-                data-cy="FilterLinkActive"
-                onClick={e => {
-                  e.preventDefault();
-                  setFilterStatus('active');
-                }}
-              >
-                Active
-              </a>
-
-              <a
-                href="#/completed"
-                className={`filter__link ${filterStatus === 'completed' ? 'selected' : ''}`}
-                data-cy="FilterLinkCompleted"
-                onClick={e => {
-                  e.preventDefault();
-                  setFilterStatus('completed');
-                }}
-              >
-                Completed
-              </a>
-            </nav>
+            <TodoFilter
+              filterStatus={filterStatus}
+              onChange={setFilterStatus}
+            />
 
             {/* this button should be disabled if there are no completed todos */}
             <button
@@ -201,21 +116,7 @@ export const App: React.FC = () => {
           </footer>
         )}
       </div>
-      {/* DON'T use conditional rendering to hide the notification */}
-      {/* Add the 'hidden' class to hide the message smoothly */}
-      <div
-        data-cy="ErrorNotification"
-        className={`notification is-danger is-light has-text-weight-normal ${!error ? 'hidden' : ''}`}
-      >
-        <button
-          data-cy="HideErrorButton"
-          type="button"
-          className="delete"
-          onClick={hideErrorNotification}
-        />
-        {/* show only one message at a time */}
-        {error}
-      </div>
+      <ErrorNotification onClose={hideErrorNotification} error={error} />
     </div>
   );
 };
