+++ src/components/ErrorNotification.tsx
@@ -1,0 +1,24 @@
+import React from 'react';
+
+type Props = {
+  error: string;
+  onClose: () => void;
+};
+
+export const ErrorNotification: React.FC<Props> = ({ error, onClose }) => {
+  return (
+    <div
+      data-cy="ErrorNotification"
+      className={`notification is-danger is-light has-text-weight-normal ${!error ? 'hidden' : ''}`}
+    >
+      <button
+        data-cy="HideErrorButton"
+        type="button"
+        className="delete"
+        onClick={onClose}
+      />
+      {/* show only one message at a time */}
+      {error}
+    </div>
+  );
+};
