+++ src/types/FilterStatus.ts
@@ -1,0 +1,5 @@
+export enum FilterStatus {
+  All = 'all',
+  Active = 'active',
+  Completed = 'completed',
+}
